\name{bedToBins}
\alias{bedToBins}
\title{Join the nearest bins to a set of bed regions}
\usage{
  bedToBins(bed, bins)
}
\arguments{
  \item{bed}{}
}
\description{
  The function accepts in a bin file (columns: cbin, chr,
  from.coord. to.coord, count) or a bin data.frame or bin
  data.table and converts to a \code{\link{GRanges}}
  object. Returns a data.table with
}
\section{\code{binToGRanges}}{

}

