\name{bedToGRanges}
\alias{bedToGRanges}
\title{Convert a bed to GRanges}
\usage{
  bedToGRanges(bed)
}
\arguments{
  \item{bed}{}
}
\description{
  The function accepts a set of sites in a bed file (no
  header, columns: chromosome,start, end, site) or a bed
  data.frame or bed data.table and converts to a
  \code{\link{GRanges}} object.
}
\section{\code{bedToGRanges}}{

}

