\name{binToGRanges}
\alias{binToGRanges}
\title{Convert a bin to GRanges}
\usage{
  binToGRanges(bin)
}
\arguments{
  \item{bed}{}
}
\description{
  The function accepts a set of sites in a bin file
  (columns: cbin, chr, from.coord. to.coord, count) or a
  bin data.frame or bin data.table and converts to a
  \code{\link{GRanges}} object.
}
\section{\code{binToGRanges}}{

}

